LayoutTests/fast/html/script-must-not-run-when-child-is-adopted-expected.txt

Adopting a script element must not run the script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing HTMLScriptElement
PASS didRunScript is false
PASS div.parentNode is not doc.body
PASS div.parentNode is null
PASS div.ownerDocument is document

Testing SVGScriptElement
PASS didRunScript is false
PASS div.parentNode is not doc.body
PASS div.parentNode is null
PASS div.ownerDocument is document
PASS successfullyParsed is true

TEST COMPLETE
